Original commit: elastic/x-pack-elasticsearch@35b4ab58325a0617252fc52eb360855a7ab14f4c

+++ elasticsearch/x-pack/marvel/src/main/java/org/elasticsearch/marvel/agent/exporter/local/LocalExporter.java
@@ -7,25 +7,22 @@ package org.elasticsearch.marvel.agent.exporter.local;
 
 import com.carrotsearch.hppc.cursors.ObjectCursor;
 import com.carrotsearch.hppc.cursors.ObjectObjectCursor;
-import org.elasticsearch.Version;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexResponse;
 import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateRequest;
 import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateResponse;
 import org.elasticsearch.cluster.ClusterChangedEvent;
-import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.ClusterStateListener;
 import org.elasticsearch.cluster.block.ClusterBlockLevel;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
+import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.regex.Regex;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.gateway.GatewayService;
-import org.elasticsearch.marvel.MarvelSettings;
-import org.elasticsearch.marvel.MonitoredSystem;
 import org.elasticsearch.marvel.agent.exporter.ExportBulk;
 import org.elasticsearch.marvel.agent.exporter.Exporter;
 import org.elasticsearch.marvel.agent.exporter.MarvelTemplateUtils;
@@ -39,6 +36,7 @@ import org.joda.time.DateTimeZone;
 
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -56,8 +54,7 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
     private final ResolversRegistry resolvers;
     private final CleanerService cleanerService;
 
-    private volatile LocalBulk bulk;
-    private volatile boolean active = true;
+    private final AtomicReference<State> state = new AtomicReference<>(State.INITIALIZED);
 
     /** Version number of built-in templates **/
     private final Integer templateVersion;
@@ -76,13 +73,8 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
         }
 
         resolvers = new ResolversRegistry(config.settings());
-        bulk = resolveBulk(clusterService.state(), bulk);
         clusterService.add(this);
         cleanerService.add(this);
-    }
-
-    LocalBulk getBulk() {
-        return bulk;
     }
 
     ResolversRegistry getResolvers() {
@@ -91,57 +83,37 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
 
     @Override
     public void clusterChanged(ClusterChangedEvent event) {
-        LocalBulk currentBulk = bulk;
-        LocalBulk newBulk = resolveBulk(event.state(), currentBulk);
-
-        // yes, this method will always be called by the cluster event loop thread
-        // but we need to sync with the {@code #close()} mechanism
-        synchronized (this) {
-            if (active) {
-                bulk = newBulk;
-            } else if (newBulk != null) {
-                newBulk.terminate();
-            }
-            if (currentBulk == null && bulk != null) {
-                logger.debug("local exporter [{}] - started!", name());
-            }
-            if (bulk != currentBulk && currentBulk != null) {
-                logger.debug("local exporter [{}] - stopped!", name());
-                currentBulk.terminate();
-            }
+        if (state.get() == State.INITIALIZED) {
+            resolveBulk(event.state());
         }
     }
 
     @Override
     public ExportBulk openBulk() {
-        return bulk;
+        if (state.get() !=  State.RUNNING) {
+            return null;
+        }
+        return resolveBulk(clusterService.state());
     }
 
-    // requires synchronization due to cluster state update events (see above)
     @Override
-    public synchronized void close() {
-        active = false;
-        clusterService.remove(this);
-        cleanerService.remove(this);
-        if (bulk != null) {
-            try {
-                bulk.terminate();
-                bulk = null;
-            } catch (Exception e) {
-                logger.error("local exporter [{}] - failed to cleanly close bulk", e, name());
-            }
+    public void doClose() {
+        if (state.getAndSet(State.TERMINATED) != State.TERMINATED) {
+            logger.debug("stopped");
+            clusterService.remove(this);
+            cleanerService.remove(this);
         }
     }
 
-    LocalBulk resolveBulk(ClusterState clusterState, LocalBulk currentBulk) {
+    LocalBulk resolveBulk(ClusterState clusterState) {
         if (clusterService.localNode() == null || clusterState == null) {
-            return currentBulk;
+            return null;
         }
 
         if (clusterState.blocks().hasGlobalBlock(GatewayService.STATE_NOT_RECOVERED_BLOCK)) {
             // wait until the gateway has recovered from disk, otherwise we think may not have .monitoring-es-
             // indices but they may not have been restored from the cluster state on disk
-            logger.debug("local exporter [{}] - waiting until gateway has recovered from disk", name());
+            logger.debug("waiting until gateway has recovered from disk");
             return null;
         }
 
@@ -151,47 +123,50 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
         // if this is not the master, we'll just look to see if the monitoring timestamped template is already
         // installed and if so, if it has a compatible version. If it is (installed and compatible)
         // we'll be able to start this exporter. Otherwise, we'll just wait for a new cluster state.
-        if (!clusterService.localNode().masterNode()) {
+        if (clusterService.state().nodes().isLocalNodeElectedMaster() == false) {
             // We only need to check the index template for timestamped indices
-            if (!templateInstalled) {
+            if (templateInstalled == false) {
                 // the template for timestamped indices is not yet installed in the given cluster state, we'll wait.
-                logger.debug("local exporter [{}] - monitoring index template does not exist, so service cannot start", name());
+                logger.debug("monitoring index template does not exist, so service cannot start");
                 return null;
             }
 
-            // ok.. we have a compatible template... we can start
-            logger.debug("local exporter [{}] - started!", name());
-            return currentBulk != null ? currentBulk : new LocalBulk(name(), logger, client, resolvers);
+            logger.debug("monitoring index template found, service can start");
+
+        } else {
+
+            // we are on master
+            //
+            // Check that there is nothing that could block metadata updates
+            if (clusterState.blocks().hasGlobalBlock(ClusterBlockLevel.METADATA_WRITE)) {
+                logger.debug("waiting until metadata writes are unblocked");
+                return null;
+            }
+
+            // Install the index template for timestamped indices first, so that other nodes can ship data
+            if (templateInstalled == false) {
+                logger.debug("could not find existing monitoring template for timestamped indices, installing a new one");
+                putTemplate(templateName, MarvelTemplateUtils.loadTimestampedIndexTemplate());
+                // we'll get that template on the next cluster state update
+                return null;
+            }
+
+            // Install the index template for data index
+            templateName = MarvelTemplateUtils.dataTemplateName(templateVersion);
+            if (hasTemplate(templateName, clusterState) == false) {
+                logger.debug("could not find existing monitoring template for data index, installing a new one");
+                putTemplate(templateName, MarvelTemplateUtils.loadDataIndexTemplate());
+                // we'll get that template on the next cluster state update
+                return null;
+            }
+
+            logger.debug("monitoring index template found on master node, service can start");
         }
 
-        // we are on master
-        //
-        // Check that there is nothing that could block metadata updates
-        if (clusterState.blocks().hasGlobalBlock(ClusterBlockLevel.METADATA_WRITE)) {
-            logger.debug("local exporter [{}] - waiting until metadata writes are unblocked", name());
-            return null;
+        if (state.compareAndSet(State.INITIALIZED, State.RUNNING)) {
+            logger.debug("started");
         }
-
-        // Install the index template for timestamped indices first, so that other nodes can ship data
-        if (!templateInstalled) {
-            logger.debug("local exporter [{}] - could not find existing monitoring template for timestamped indices, installing a new one",
-                    name());
-            putTemplate(templateName, MarvelTemplateUtils.loadTimestampedIndexTemplate());
-            // we'll get that template on the next cluster state update
-            return null;
-        }
-
-        // Install the index template for data index
-        templateName = MarvelTemplateUtils.dataTemplateName(templateVersion);
-        if (!hasTemplate(templateName, clusterState)) {
-            logger.debug("local exporter [{}] - could not find existing monitoring template for data index, installing a new one", name());
-            putTemplate(templateName, MarvelTemplateUtils.loadDataIndexTemplate());
-            // we'll get that template on the next cluster state update
-            return null;
-        }
-
-        // ok.. we have a compatible templates... we can start
-        return currentBulk != null ? currentBulk : new LocalBulk(name(), logger, client, resolvers);
+        return new LocalBulk(name(), logger, client, resolvers);
     }
 
     /**
@@ -208,7 +183,7 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
                 try {
                     Integer version = Integer.parseInt(template.value.substring(templatePattern.length() - 1));
                     templates.put(template.value, version);
-                    logger.debug("found index template [{}] in version [{}]", template, version);
+                    logger.debug("found index template [{}] in version [{}]", template.value, version);
                 } catch (NumberFormatException e) {
                     logger.warn("cannot extract version number for template [{}]", template.value);
                 }
@@ -223,7 +198,7 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
     }
 
     void putTemplate(String template, byte[] source) {
-        logger.debug("local exporter [{}] - installing template [{}]", name(), template);
+        logger.debug("installing template [{}]",template);
 
         PutIndexTemplateRequest request = new PutIndexTemplateRequest(template).source(source);
         assert !Thread.currentThread().isInterrupted() : "current thread has been interrupted before putting index template!!!";
@@ -233,31 +208,31 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
             @Override
             public void onResponse(PutIndexTemplateResponse response) {
                 if (response.isAcknowledged()) {
-                    logger.trace("local exporter [{}] - successfully installed monitoring template [{}]", name(), template);
+                    logger.trace("successfully installed monitoring template [{}]", template);
                 } else {
-                    logger.error("local exporter [{}] - failed to update monitoring index template [{}]", name(), template);
+                    logger.error("failed to update monitoring index template [{}]", template);
                 }
             }
 
             @Override
             public void onFailure(Throwable throwable) {
-                logger.error("local exporter [{}] - failed to update monitoring index template [{}]", throwable, name(), template);
+                logger.error("failed to update monitoring index template [{}]", throwable, template);
             }
         });
     }
 
     @Override
     public void onCleanUpIndices(TimeValue retention) {
-        if (bulk == null) {
-            logger.debug("local exporter [{}] - not ready yet", name());
+        if (state.get() != State.RUNNING) {
+            logger.debug("exporter not ready");
             return;
         }
 
-        if (clusterService.localNode().masterNode()) {
+        if (clusterService.state().nodes().isLocalNodeElectedMaster()) {
             // Reference date time will be compared to index.creation_date settings,
             // that's why it must be in UTC
             DateTime expiration = new DateTime(DateTimeZone.UTC).minus(retention.millis());
-            logger.debug("local exporter [{}] - cleaning indices [expiration={}, retention={}]", name(), expiration, retention);
+            logger.debug("cleaning indices [expiration={}, retention={}]", expiration, retention);
 
             ClusterState clusterState = clusterService.state();
             if (clusterState != null) {
@@ -291,7 +266,7 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
                         long creationDate = index.value.getCreationDate();
                         if (creationDate <= expirationTime) {
                             if (logger.isDebugEnabled()) {
-                                logger.debug("local exporter [{}] - detected expired index [name={}, created={}, expired={}]", name(),
+                                logger.debug("detected expired index [name={}, created={}, expired={}]",
                                         indexName, new DateTime(creationDate, DateTimeZone.UTC), expiration);
                             }
                             indices.add(indexName);
@@ -300,31 +275,33 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
                 }
 
                 if (!indices.isEmpty()) {
-                    logger.info("local exporter [{}] - cleaning up [{}] old indices", name(), indices.size());
+                    logger.info("cleaning up [{}] old indices", indices.size());
                     deleteIndices(indices);
                 } else {
-                    logger.debug("local exporter [{}] - no old indices found for clean up", name());
+                    logger.debug("no old indices found for clean up");
                 }
             }
         }
     }
 
     private void deleteIndices(Set<String> indices) {
-        logger.trace("local exporter [{}] - deleting {} indices: {}", name(), indices.size(), collectionToCommaDelimitedString(indices));
+        logger.trace("deleting {} indices: [{}]", indices.size(), collectionToCommaDelimitedString(indices));
         client.admin().indices().delete(new DeleteIndexRequest(indices.toArray(new String[indices.size()])),
                 new ActionListener<DeleteIndexResponse>() {
             @Override
             public void onResponse(DeleteIndexResponse response) {
                 if (response.isAcknowledged()) {
-                    logger.debug("local exporter [{}] - indices deleted", name());
+                    logger.debug("{} indices deleted", indices.size());
                 } else {
-                    logger.warn("local exporter [{}] - unable to delete {} indices", name(), indices.size());
+                    // Probably means that the delete request has timed out,
+                    // the indices will survive until the next clean up.
+                    logger.warn("deletion of {} indices wasn't acknowledged", indices.size());
                 }
             }
 
             @Override
             public void onFailure(Throwable e) {
-                logger.error("local exporter [{}] - failed to delete indices", e, name());
+                logger.error("failed to delete indices", e);
             }
         });
     }
@@ -347,5 +324,11 @@ public class LocalExporter extends Exporter implements ClusterStateListener, Cle
         public LocalExporter create(Config config) {
             return new LocalExporter(config, client, clusterService, cleanerService);
         }
+    }
+
+    enum State {
+        INITIALIZED,
+        RUNNING,
+        TERMINATED
     }
 }
